backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-dotenv>=1.0.1
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.9

# Database - Cloud SQL PostgreSQL
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
cloud-sql-python-connector[asyncpg]>=1.0.0
